改id方案，修复打开浏览器bug

+++ dao/category.go
@@ -9,11 +9,11 @@ func AddCategory(db *gorm.DB, category model.Category) (model.Category, error) {
 	return category, db.Create(&category).Error
 }
 
-func GetCategory(db *gorm.DB, id int) (category model.Category, err error) {
+func GetCategory(db *gorm.DB, id int64) (category model.Category, err error) {
 	return category, db.Where("id=?", id).Find(&category).Error
 }
 
-func GetCategories(db *gorm.DB, pid int) (categories []model.Category, err error) {
+func GetCategories(db *gorm.DB, pid int64) (categories []model.Category, err error) {
 	return categories, db.Where("pid=?", pid).Find(&categories).Error
 }
